requirements.txt
boto3==1.4.7
pyyaml==3.12
